devpulse_backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
feedparser==6.0.11
idna==3.10
praw==7.8.1
prawcore==2.4.0
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
sgmllib3k==1.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
update-checker==0.18.0
urllib3==2.4.0
websocket-client==1.8.0
